% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktioner vedr tekststrenge.R
\encoding{UTF-8}
\name{foersteBogstavTilLille}
\alias{foersteBogstavTilLille}
\title{Ændre første bogstav i tekst til lille bogstav}
\usage{
foersteBogstavTilLille(x)
}
\arguments{
\item{x}{en tekststreng}
}
\value{
teksten x med lille begyndelsesbogstav
}
\description{
Ændre første bogstav i tekst til lille bogstav
}
\examples{
foersteBogstavTilLille("Hej")
}
