% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktioner vedr tekststrenge.R
\encoding{UTF-8}
\name{foersteBogstavTilStor}
\alias{foersteBogstavTilStor}
\title{Ændre første bogstav i tekst til stort bogstav}
\usage{
foersteBogstavTilStor(x)
}
\arguments{
\item{x}{en tekststreng}
}
\value{
teksten x med stort begyndelsesbogstav
}
\description{
Ændre første bogstav i tekst til stort bogstav
}
\examples{
foersteBogstavTilStor("hej")
}
