% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktioner vedr tekststrenge.R
\encoding{UTF-8}
\name{opdelCamelCaseTekst}
\alias{opdelCamelCaseTekst}
\title{Opdel (lower) Camel Case tekst til tekst med mellemrum}
\usage{
opdelCamelCaseTekst(tekst)
}
\arguments{
\item{tekst}{en tekst af typen lowerCamelCase eller UpperCamelCase.}
}
\value{
'tekst' opdelt med mellemrum
}
\description{
Opdel (lower) Camel Case tekst til tekst med mellemrum
}
\examples{
opdelCamelCaseTekst("lowerCamelCase")
}
