% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktioner vedr tal.R
\encoding{UTF-8}
\name{procent}
\alias{procent}
\title{Procent funktion}
\usage{
procent(taeller, naevner, antalDecimal = 2)
}
\arguments{
\item{taeller}{et tal der repræsenterer tælleren}

\item{naevner}{et tal der repræsenterer nævneren}

\item{antalDecimal}{det ønskede antal decimaler i resultatet. Default er sat til to decimaler.}
}
\value{
procent med '%'-tegn
}
\description{
Funktion til at beregne procent og returnere det med '%' tegn
}
\examples{
procent(taeller = 25, naevner = 50, antalDecimal = 0) eller procent(25, 50, 0)
}
