% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktioner vedr tekststrenge.R
\encoding{UTF-8}
\name{retBindestregTekstTilLCC}
\alias{retBindestregTekstTilLCC}
\title{Ret tekst med bindestreg opdeling til lowerCamelCase}
\usage{
retBindestregTekstTilLCC(tekst)
}
\arguments{
\item{tekst}{en chracter string, teksten med bindestreg-opdeling}
}
\value{
'tekst' i lowerCamelCase form
}
\description{
Ret tekst med bindestreg opdeling til lowerCamelCase
}
