% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktioner vedr tekststrenge.R
\encoding{UTF-8}
\name{retTekstTilLCC}
\alias{retTekstTilLCC}
\title{Ret tekst til lowerCamelCase}
\usage{
retTekstTilLCC(tekst)
}
\arguments{
\item{tekst}{en character string, teksten}
}
\value{
'tekst' i lowerCamelCase form
}
\description{
Ret tekst til lowerCamelCase
}
