% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktioner vedr tal.R
\encoding{UTF-8}
\name{rundTilNaermesteY}
\alias{rundTilNaermesteY}
\title{Rund tallet til nærmeste Y}
\usage{
rundTilNaermesteY(x, y)
}
\arguments{
\item{x}{et tal man øsnker at runde til nærmeste Y}

\item{y}{et tal der definere det nærmeste tal-interval man vil runde op/ned til}
}
\value{
x rundet op/ned til den nærmeste multiplum af Y
}
\description{
Rund tallet til nærmeste Y
}
\examples{
rundTilNaermesteY(21, 5)
}
