% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktioner vedr tal.R
\encoding{UTF-8}
\name{talMedTusindtalssep}
\alias{talMedTusindtalssep}
\title{Tal med tusindtalsseparator}
\usage{
talMedTusindtalssep(x)
}
\arguments{
\item{x}{Et tal som man ønsker returneret med tusindtalsseparator}
}
\value{
X med tusindtal (,) og decimalseparator (.)
}
\description{
Tal med tusindtalsseparator
}
\examples{
talFormat(1000.10) vil returnere 1,000.10
}
